Revert last bad commit
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40336 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSDragView.h
@@ -129,7 +129,7 @@
 		 timestamp: (NSTimeInterval)time
 		  toWindow: (NSInteger)dWindowNumber;
 - (NSWindow*) windowAcceptingDnDunder: (NSPoint)mouseLocation
-			    windowRef: (NSInteger*)mouseWindowRef;
+			    windowRef: (int*)mouseWindowRef;
 
 @end
 #endif
